This version is a WIP and is kinda weird.
At startup it takes an array of data, a BMP converted to a C array at https://lvgl.io/tools/imageconverter, 2 indexed colors (0 is black, 1 is white, so the data has to be inverted), and repacks it into 8-bit vertical lines for the Sharp PC-G850 to display. It produces BASIC code that is pasted into the yexy area. This will move later to accept a file, and produce the code on demand. For now, bear with me.

LoRa.asm introduces MX2KEY, a method to convert the key matrix to an ASCII char (if possible). Useful when calling WAITK (BFCDH):

```
470MAIN0: CALL WAITK
515 CALL MX2KEY
480 CP 'Q' ; it would be 02H normally
490 JP Z, THEEND
500 CP 'R' ; it would be 05H normally
510 JP Z, RECV
```