I just acquired a PC-E200, the first of that line of pokekon, so things are a little bit different, and need adjusting. No assembler on the device, and I am having difficultie (as in No way José) receiving binhex data onto the device. W works, R doesn't...
So I needed a way to have a toolchain: I know I could install SDCC,
but the idea is to keep the code exactly as is. So I went coding a
monster. Here's the process.

1. Compile on the PC-G850(V/S).
2. Export the compiled code via W in the Monitor.
3. Convert the binhex to BASIC code that pokes the binary code.
4. Upload the BASIC code to the PC-E200, and run it.
Voilà, you have the binary in place.
The conversion to BASIC code is done in Xojo, I'll put up a project later.