Update ASM.md
Tried this Rom routine but returned something else than what was described so I searched the manual and found it does a little bit more. I'm only getting in the PC-G850 and you have an excellent resource and I am actively using it, so hope you allow me to update this as others may benefit.